Extracted a base definition for the Clonk
In external packs a lot of times you do not need or want all of the
Clonk behaviour. This is a first step towards making it more modular.
