+++ planet/Objects.ocd/Clonk.ocd/Generic.ocd/StringTblDE.txt
@@ -1,0 +1,7 @@
+SetMaxContentsCount=Inventarplatzzahl aendern
+SetMaxContentsCountHelp=Aendert die maximale Anzahl an Objekten, die der Clonk tragen kann. Gueltige Werte von 0 bis 10. Ein zusaetzliches CarryHeavy-Objekt kann immer getragen werden.
+MaxContentsCount=Anzahl Inventarplaetze
+Default=Standard
+
+TurnAround=Umdrehen
+TurnAroundHelp=Wechselt die Blickrichtung des Clonks.
