+++ planet/Objects.ocd/Clonk.ocd/Generic.ocd/StringTblUS.txt
@@ -1,0 +1,7 @@
+SetMaxContentsCount=Change inventory size
+SetMaxContentsCountHelp=Modifies the maximum number of items a clonk can carry. Valid values range from 0 to 10. One additional carry-heavy object can always be carried.
+MaxContentsCount=Inventory slot count
+Default=Default
+
+TurnAround=Turn around
+TurnAroundHelp=Changes the view direction of the clonk.
