+++ planet/Objects.ocd/Clonk.ocd/StringTblDE.txt
@@ -11,12 +11,5 @@ TxtEnergysupply=Energieversorgung
 
 MsgDeadClonk=%s ist tot.|Es sind keine Interaktionen möglich.
 
-SetMaxContentsCount=Inventarplatzzahl aendern
-SetMaxContentsCountHelp=Aendert die maximale Anzahl an Objekten, die der Clonk tragen kann. Gueltige Werte von 0 bis 10. Ein zusaetzliches CarryHeavy-Objekt kann immer getragen werden.
-MaxContentsCount=Anzahl Inventarplaetze
-Default=Standard
-
 Skin=Skin
 SkinHelp=Aussehen des Clonks.
-TurnAround=Umdrehen
-TurnAroundHelp=Wechselt die Blickrichtung des Clonks.
